[#1714] Remove file lock abstraction and put this directly into filedescriptor management
