+++ iceoryx2-bb/posix/src/process_state.rs
@@ -154,8 +154,7 @@ use crate::{
     access_mode::AccessMode,
     directory::{Directory, DirectoryAccessError, DirectoryCreateError},
     file::{File, FileBuilder, FileCreationError, FileOpenError, FileRemoveError, FileStatError},
-    file_descriptor::{FileDescriptorBased, FileDescriptorManagement},
-    file_lock::LockType,
+    file_descriptor::{FileDescriptorBased, FileDescriptorManagement, LockType},
     permission::Permission,
     process::{Process, UniqueProcessId},
     unix_datagram_socket::CreationMode,
